refactor: Перенос transport_moc_solver в отдельный файл
